on payment card update added exception handling

+++ app/Http/Controllers/API/PaymentCardDetailController.php
@@ -190,10 +190,22 @@ class PaymentCardDetailController extends Controller
 
             if(!empty($paymentCardDetail->stripe_payment_method_id))
             {
-                $stripe->paymentMethods->detach(
-                    $paymentCardDetail->stripe_payment_method_id,
-                    []
-                );
+                $e = null;
+                try {
+                    $getInfo = $stripe->paymentMethods->retrieve(
+                      $paymentCardDetail->stripe_payment_method_id,
+                      []
+                    );
+                } catch (\Exception $e) {
+                    
+                }
+                if(empty($e))
+                {
+                    $stripe->paymentMethods->detach(
+                        $paymentCardDetail->stripe_payment_method_id,
+                        []
+                    );
+                }
             }
 
             $paymentMethods = $stripe->paymentMethods->create([
